+++ src/webportal/src/app/job/job-view/fabric/job-detail/components/monaco-modal.jsx
@@ -24,7 +24,7 @@ import React from 'react';
 import MonacoEditor from 'react-monaco-editor';
 
 import {monacoHack} from './monaco-hack.scss';
-import t from '../../tachyons.css';
+import t from '../../../../../../css/tachyons.css';
 
 const MonacoModal = ({isOpen, onDismiss, title, monacoProps}) => (
   <div>
