+++ src/webportal/src/app/job/job-view/fabric/job-detail/components/monaco-panel.jsx
@@ -25,7 +25,7 @@ import React from 'react';
 import MonacoEditor from 'react-monaco-editor';
 
 import {monacoHack} from './monaco-hack.scss';
-import t from '../../tachyons.css';
+import t from '../../../../../../css/tachyons.css';
 
 export default class MonacoPanel extends React.Component {
   constructor(props) {
