+++ src/webportal/src/app/job/job-view/fabric/job-detail/components/status-badge.jsx
@@ -22,7 +22,7 @@ import {Icon} from 'office-ui-fabric-react/lib/Icon';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import t from '../../tachyons.css';
+import t from '../../../../../../css/tachyons.css';
 
 import {statusColorMapping} from '../util';
