+++ src/webportal/src/app/job/job-view/fabric/job-detail/components/summary.jsx
@@ -26,7 +26,7 @@ import {MessageBar, MessageBarType} from 'office-ui-fabric-react/lib/MessageBar'
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import t from '../../tachyons.css';
+import t from '../../../../../../css/tachyons.css';
 
 import Card from './card';
 import MonacoPanel from './monaco-panel';
