+++ src/webportal/src/app/job/job-view/fabric/job-detail/components/top.jsx
@@ -18,7 +18,7 @@
 import React from 'react';
 import {ActionButton} from 'office-ui-fabric-react/lib/Button';
 
-import t from '../../tachyons.css';
+import t from '../../../../../../css/tachyons.css';
 
 const Top = () => (
   <div className={t.flex}>
