--- src/webportal/src/app/job/job-view/fabric/tachyons.css
@@ -1,62 +1,0 @@
-/* Modules */
-@import '../../../../../node_modules/tachyons/src/_box-sizing';
-@import '../../../../../node_modules/tachyons/src/_aspect-ratios';
-@import '../../../../../node_modules/tachyons/src/_images';
-@import '../../../../../node_modules/tachyons/src/_background-size';
-@import '../../../../../node_modules/tachyons/src/_background-position';
-@import '../../../../../node_modules/tachyons/src/_outlines';
-@import '../../../../../node_modules/tachyons/src/_borders';
-@import '../../../../../node_modules/tachyons/src/_border-colors';
-@import '../../../../../node_modules/tachyons/src/_border-radius';
-@import '../../../../../node_modules/tachyons/src/_border-style';
-@import '../../../../../node_modules/tachyons/src/_border-widths';
-@import '../../../../../node_modules/tachyons/src/_box-shadow';
-@import '../../../../../node_modules/tachyons/src/_code';
-@import '../../../../../node_modules/tachyons/src/_coordinates';
-@import '../../../../../node_modules/tachyons/src/_clears';
-@import '../../../../../node_modules/tachyons/src/_display';
-@import '../../../../../node_modules/tachyons/src/_flexbox';
-@import '../../../../../node_modules/tachyons/src/_floats';
-@import '../../../../../node_modules/tachyons/src/_font-family';
-@import '../../../../../node_modules/tachyons/src/_font-style';
-@import '../../../../../node_modules/tachyons/src/_font-weight';
-@import '../../../../../node_modules/tachyons/src/_forms';
-@import '../../../../../node_modules/tachyons/src/_heights';
-@import '../../../../../node_modules/tachyons/src/_letter-spacing';
-@import '../../../../../node_modules/tachyons/src/_line-height';
-@import '../../../../../node_modules/tachyons/src/_links';
-@import '../../../../../node_modules/tachyons/src/_lists';
-@import '../../../../../node_modules/tachyons/src/_max-widths';
-@import '../../../../../node_modules/tachyons/src/_widths';
-@import '../../../../../node_modules/tachyons/src/_overflow';
-@import '../../../../../node_modules/tachyons/src/_position';
-@import '../../../../../node_modules/tachyons/src/_opacity';
-@import '../../../../../node_modules/tachyons/src/_rotations';
-@import '../../../../../node_modules/tachyons/src/_skins';
-@import '../../../../../node_modules/tachyons/src/_skins-pseudo';
-@import '../../../../../node_modules/tachyons/src/_spacing';
-@import '../../../../../node_modules/tachyons/src/_negative-margins';
-@import '../../../../../node_modules/tachyons/src/_tables';
-@import '../../../../../node_modules/tachyons/src/_text-decoration';
-@import '../../../../../node_modules/tachyons/src/_text-align';
-@import '../../../../../node_modules/tachyons/src/_text-transform';
-@import '../../../../../node_modules/tachyons/src/_type-scale';
-@import '../../../../../node_modules/tachyons/src/_typography';
-@import '../../../../../node_modules/tachyons/src/_utilities';
-@import '../../../../../node_modules/tachyons/src/_visibility';
-@import '../../../../../node_modules/tachyons/src/_white-space';
-@import '../../../../../node_modules/tachyons/src/_vertical-align';
-@import '../../../../../node_modules/tachyons/src/_hovers';
-@import '../../../../../node_modules/tachyons/src/_z-index';
-@import '../../../../../node_modules/tachyons/src/_nested';
-@import '../../../../../node_modules/tachyons/src/_styles';
-
-/* Variables */
-/* Importing here will allow you to override any variables in the modules */
-@import '../../../../../node_modules/tachyons/src/_colors';
-@import '../../../../../node_modules/tachyons/src/_media-queries';
-
-/* Debugging */
-@import '../../../../../node_modules/tachyons/src/_debug-children';
-@import '../../../../../node_modules/tachyons/src/_debug-grid';
-
